% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EEZ_Baltic.R
\docType{data}
\name{EEZ_Baltic}
\alias{EEZ_Baltic}
\title{Shapefile}
\format{
Shapefile
}
\usage{
data(EEZ_Baltic)
}
\description{
Shapefile from EEZ_Baltic
}
\examples{
data(EEZ_Baltic)
}
\references{
To-do
}
\keyword{datasets}
