% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EEZ_North.R
\docType{data}
\name{EEZ_North}
\alias{EEZ_North}
\title{Shapefile}
\format{
Shapefile
}
\usage{
data(EEZ_North)
}
\description{
Shapefile from EEZ_North
}
\examples{
data(EEZ_North)
}
\references{
To-do
}
\keyword{datasets}
