% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/German_bath.R
\docType{data}
\name{German_bath}
\alias{German_bath}
\title{Shapefile}
\format{
Shapefile
}
\usage{
data(German_bath)
}
\description{
Shapefile from German Bathymetry
}
\examples{
data(German_bath)
}
\references{
To-do
}
\keyword{datasets}
