% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/German_coast.R
\docType{data}
\name{German_coast}
\alias{German_coast}
\title{Shapefile}
\format{
Shapefile
}
\usage{
data(German_coast)
}
\description{
Shapefile from German Coast
}
\examples{
data(German_coast)
}
\references{
To-do
}
\keyword{datasets}
