% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/German_distancecoast.R
\docType{data}
\name{German_distancecoast}
\alias{German_distancecoast}
\title{Shapefile}
\format{
Shapefile
}
\usage{
data(German_distancecoast)
}
\description{
Shapefile from German Distance to Coast
}
\examples{
data(German_distancecoast)
}
\references{
To-do
}
\keyword{datasets}
