% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/German_EEZ.R
\docType{data}
\name{German_EEZ}
\alias{German_EEZ}
\title{Shapefile}
\format{
Shapefile
}
\usage{
data(German_EEZ)
}
\description{
Shapefile from German EEZ
}
\examples{
data(German_EEZ)
}
\references{
To-do
}
\keyword{datasets}
