% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/German_ICES.R
\docType{data}
\name{German_ICES}
\alias{German_ICES}
\title{Shapefile}
\format{
Shapefile
}
\usage{
data(German_ICES)
}
\description{
Shapefile from German_ICES
}
\examples{
data(German_ICES)
}
\references{
To-do
}
\keyword{datasets}
