% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/German_land.R
\docType{data}
\name{German_land}
\alias{German_land}
\title{Shapefile}
\format{
Shapefile
}
\usage{
data(German_land)
}
\description{
Shapefile from German land
}
\examples{
data(German_land)
}
\references{
To-do
}
\keyword{datasets}
