% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/German_natura.R
\docType{data}
\name{German_natura}
\alias{German_natura}
\title{Shapefile}
\format{
Shapefile
}
\usage{
data(German_natura)
}
\description{
Shapefile from German natura
}
\examples{
data(German_natura)
}
\references{
To-do
}
\keyword{datasets}
