% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/German_OWF.R
\docType{data}
\name{German_OWF}
\alias{German_OWF}
\title{Shapefile}
\format{
Shapefile
}
\usage{
data(German_OWF)
}
\description{
Shapefile from German_OWF
}
\examples{
data(German_OWF)
}
\references{
To-do
}
\keyword{datasets}
