% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/German_SCA.R
\docType{data}
\name{German_SCA}
\alias{German_SCA}
\title{Shapefile}
\format{
Shapefile
}
\usage{
data(German_SCA)
}
\description{
Shapefile from German Special Conservation Areas
}
\examples{
data(German_SCA)
}
\references{
To-do
}
\keyword{datasets}
