% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/German_Shipping.R
\docType{data}
\name{German_Shipping}
\alias{German_Shipping}
\title{Shapefile}
\format{
Shapefile
}
\usage{
data(German_Shipping)
}
\description{
Shapefile from German_Shipping
}
\examples{
data(German_Shipping)
}
\references{
To-do
}
\keyword{datasets}
