% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Germany.R
\docType{data}
\name{Germany}
\alias{Germany}
\title{Shapefile}
\format{
Shapefile
}
\usage{
data(Germany)
}
\description{
Shapefile from Germany
}
\examples{
data(Germany)
}
\references{
To-do
}
\keyword{datasets}
