% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid10x10_3035.R
\docType{data}
\name{grid10x10_3035}
\alias{grid10x10_3035}
\title{Contains a sf object with a grid from the North Sea of 10 x 10
A total of 1424 observations of 2 variables}
\format{
A grid 10x10 of the north sea
}
\usage{
data(grid10x10_3035)
}
\description{
Contains a sf object with a grid from the North Sea of 10 x 10
A total of 1424 observations of 2 variables
}
\examples{
data(grid10x10_3035)
}
\references{
Lerma et al. 2023
}
\keyword{datasets}
