% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid5x5_3035.R
\docType{data}
\name{grid5x5_3035}
\alias{grid5x5_3035}
\title{Contains a sf object with a grid from the North Sea of 5 x 5
A total of 4080 observations of 2 variables}
\format{
A grid 5x5 of the north sea
}
\usage{
data(grid5x5_3035)
}
\description{
Contains a sf object with a grid from the North Sea of 5 x 5
A total of 4080 observations of 2 variables
}
\examples{
data(grid5x5_3035)
}
\references{
Lerma et al. 2023
}
\keyword{datasets}
