% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWF_EMODnet.R
\docType{data}
\name{OWF_EMODnet}
\alias{OWF_EMODnet}
\title{Shapefile}
\format{
Shapefile
}
\usage{
data(OWF_EMODnet)
}
\description{
Shapefile from German EMODnet
}
\examples{
data(OWF_EMODnet)
}
\references{
To-do
}
\keyword{datasets}
