% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study_areas.R
\docType{data}
\name{Study_areas}
\alias{Study_areas}
\title{Shapefile}
\format{
Shapefile
}
\usage{
data(Study_areas)
}
\description{
Shapefile from Study_areas BfN
}
\examples{
data(Study_areas)
}
\references{
Bundesamt für Naturschutz (BfN)
}
\keyword{datasets}
